requirements.txt
gtts
python-docx
PyPDF2
PyPDF2[crypto]
pytest